Scraper CLI final product first draft
