+++ src/scraper/main.ts
@@ -6,64 +6,65 @@ import path from "node:path";
 import { insertCourseData } from "./db";
 import { termsT } from "@/db/schema";
 import { eq } from "drizzle-orm";
+import { TermScrape } from "./types";
 
 const CACHE_PATH = "cache/";
 // const TERMS = ["202610", "202530", "202534", "202532"];
 const TERMS = ["202610"];
+const args = process.argv.splice(2);
 
-// always assume this file is only every called directy
 (async () => {
-  for (const term of TERMS) {
-    await main(term);
-  }
-})();
+    for (const term of TERMS) {
+      const cachedTerm = await cacheTerm(term);
 
-async function main(m: string) {
-  const cachename = path.resolve(CACHE_PATH, `term-${m}.json`);
-  const existingCache = existsSync(cachename);
+      if (args.includes("--dbpush")){
+        await dbPush(term, cachedTerm)
+      }
+    }
+  })();
 
-  let term;
-  if (existingCache) {
-    console.log("existing cache found");
-    term = JSON.parse(readFileSync(cachename, "utf8"));
-  } else {
-    console.log("generating new scrape");
-    term = await scrapeTerm(m);
+//No flags: prioritizes data in cache; if no data in cache, re-scrapes per missing term.
+//--fresh flag: re-creates the entire cache, regardless of prior data existed.
+async function cacheTerm(termStr: string) {
+    const cachename = path.resolve(CACHE_PATH, `term-${termStr}.json`);
+    const existingCache = existsSync(cachename);
 
-    writeFile(cachename, JSON.stringify(term), (err) => {
-      if (err) console.log(err);
+    let term
+    if (args.includes("--fresh") || !existingCache) {
+        console.log("generating new scrape");
+        term = await scrapeTerm(termStr);
+
+        writeFile(cachename, JSON.stringify(term), (err) => {
+        if (err) console.log(err);
+        });
+    } else {
+        console.log("existing cache found");
+        term = JSON.parse(readFileSync(cachename, "utf8"));
+    }
+    return term
+}
+//if we want to push, use the flag --dbpush
+async function dbPush(term: string, cachedTerm: TermScrape) {
+    //Database Shenanigans
+    const projectDir = process.cwd();
+    loadEnvConfig(projectDir);
+    const db = drizzle({
+        connection: process.env.DATABASE_URL_DIRECT!,
     });
-  }
 
-  const projectDir = process.cwd();
-  loadEnvConfig(projectDir);
-  const db = drizzle({
-    connection: process.env.DATABASE_URL_DIRECT!,
-  });
+    console.log("connected");
 
-  console.log("connected");
+    const existingTerm = await db
+        .select({})
+        .from(termsT)
+        .where(eq(termsT.term, term));
 
-  const existingTerm = await db
-    .select({})
-    .from(termsT)
-    .where(eq(termsT.term, m));
+    if (existingTerm.length > 0) {
+        console.log("term already exists... skipping");
+        console.log("paritally updating terms maybe coming?");
+        console.log("it needs to happen but that's so much work rn lowkey");
+        return;
+    }
 
-  if (existingTerm.length > 0) {
-    console.log("term already exists... skipping");
-    console.log("paritally updating terms maybe coming?");
-    console.log("it needs to happen but that's so much work rn lowkey");
-    return;
-  }
-
-  await insertCourseData(term, db);
-
-  // generate the searching index
-  // BUG: this is being a little problematic - really we should just drop and reindex completely
-  //   await db.execute(sql`
-  //     CREATE INDEX IF NOT EXISTS courses_search_idx ON courses
-  //     USING bm25 (id, name, subject, "courseNumber")
-  //     WITH (key_field='id',
-  //         text_fields='{"name": {"tokenizer": {"type": "ngram", "min_gram": 4, "max_gram": 5, "prefix_only": false}}}'
-  //     );
-  // `);
+    await insertCourseData(cachedTerm, db);
 }
